Fixes Issue with Newlines between nested lists.
Basically the issue is that when you have a nested list:

    - Item
      - Item 1
      - Item 2

And you save that in Geeknote. When you open it again you get:

    - Item

      - Item 1
      - Item 2

Note the separation between the two.

The reason for this to happen, is that Evernote adds a <p> inside the li that
has a nested <ul> element. The above commit just removes that <p> so that
html2text outputs some correct markdown.

Conflicts:
	geeknote/editor.py