Fixed resources is not defined bug in _parseInput
Fix for error on in the _parseInput function of geeknote.py

Python was throwing an error due to name resources being evaluated
before it was defined.  Someone more familiar with the code base of
geeknote can likely do a better job and make sure that the resources var
is defined (even if None).  For now I just check (correctly) that
'resources' lives in vars().