oauth.py: Add HTTP proxy support
This patch adds support for a HTTP proxy to the geeknote oauth
authorization code. The proxy is read from system settings,
usually the environment variable https_proxy.
Basic proxy authorization is supported.

Note that for geeknote to work through a proxy, proxy support
for thrift needs to be enabled, too. See
https://issues.apache.org/jira/browse/THRIFT-3798